점프 구현 및 BoxCollider엔진으로 옮기기.

+++ Engine/Render/Renderer.cpp
@@ -6,7 +6,7 @@ namespace Wanted
 {
 	Renderer::Frame::Frame(int bufferCount)
 	{
-		// ¹è¿­ »ý¼º ¹× ÃÊ±âÈ­.
+		// ï¿½è¿­ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ ï¿½Ê±ï¿½È­.
 		charInfoArray = new CHAR_INFO[bufferCount];
 		memset(charInfoArray, 0, sizeof(CHAR_INFO) * bufferCount);
 
@@ -22,23 +22,23 @@ namespace Wanted
 
 	void Renderer::Frame::Clear(const Vector2 screenSize)
 	{
-		// 2Â÷¿ø ¹è¿­·Î ´Ù·ç´Â 1Â÷¿ø ¹è¿­À» ¼øÈ¸ÇÏ¸é¼­ ºó ¹®ÀÚ¸¦ ¼³Á¤.
-		const int width = screenSize.x;
-		const int height = screenSize.y;
+		// 2ï¿½ï¿½ï¿½ï¿½ ï¿½è¿­ï¿½ï¿½ ï¿½Ù·ï¿½ï¿½ 1ï¿½ï¿½ï¿½ï¿½ ï¿½è¿­ï¿½ï¿½ ï¿½ï¿½È¸ï¿½Ï¸é¼­ ï¿½ï¿½ ï¿½ï¿½ï¿½Ú¸ï¿½ ï¿½ï¿½ï¿½ï¿½.
+		const int width = static_cast<int>(screenSize.x);
+		const int height = static_cast<int>(screenSize.y);
 
 		for (int y = 0; y < height; y++)
 		{
 			for (int x = 0; x < width; x++)
 			{
-				// ¹è¿­ ÀÎµ¦½º ±¸ÇÏ±â.
+				// ï¿½è¿­ ï¿½Îµï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½Ï±ï¿½.
 				const int index = (y * width) + x;
 
-				// ±ÛÀÚ °ª ¹× ¼Ó¼º ¼³Á¤.
+				// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ ï¿½ï¿½ ï¿½Ó¼ï¿½ ï¿½ï¿½ï¿½ï¿½.
 				CHAR_INFO& info = charInfoArray[index];
 				info.Char.AsciiChar = ' ';
 				info.Attributes = 0;
 
-				// ±×¸®±â ¿ì¼±¼øÀ§ ÃÊ±âÈ­.
+				// ï¿½×¸ï¿½ï¿½ï¿½ ï¿½ì¼±ï¿½ï¿½ï¿½ï¿½ ï¿½Ê±ï¿½È­.
 				sortingOrderArray[index] = -1;
 			}
 		}
@@ -59,21 +59,21 @@ namespace Wanted
 	{
 		instance = this;
 
-		// ÇÁ·¹ÀÓ °´Ã¼ »ý¼º.
-		const int bufferCount = screenSize.x * screenSize.y;
+		// ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½Ã¼ ï¿½ï¿½ï¿½ï¿½.
+		const int bufferCount = static_cast<int>(screenSize.x * screenSize.y);
 		frame = new Frame(bufferCount);
 
-		// ÇÁ·¹ÀÓ ÃÊ±âÈ­.
+		// ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½Ê±ï¿½È­.
 		frame->Clear(screenSize);
 
-		// ÀÌÁß ¹öÆÛ °´Ã¼ »ý¼º.
+		// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½Ã¼ ï¿½ï¿½ï¿½ï¿½.
 		screenBuffers[0] = new ScreenBuffer(screenSize);
 		screenBuffers[0]->Clear();
 
 		screenBuffers[1] = new ScreenBuffer(screenSize);
 		screenBuffers[1]->Clear();
 
-		// È°¼ºÈ­ ¹öÆÛ ¼³Á¤.
+		// È°ï¿½ï¿½È­ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½.
 		Present();
 	}
 
@@ -99,98 +99,98 @@ namespace Wanted
 
 	void Renderer::Draw()
 	{
-		// È­¸é Áö¿ì±â.
+		// È­ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½.
 		Clear();
 
-		// ÀüÁ¦Á¶°Ç : ·¹º§ÀÇ ¸ðµç ¾×ÅÍ°¡ ·»´õ·¯¿¡ SubmitÀ» ¿Ï·á.
-		// ·»´õÅ¥ ¼øÈ¸ÇÏ¸é¼­ ÇÁ·¹ÀÓ Ã¤¿ì±â.
+		// ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ : ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½Í°ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ Submitï¿½ï¿½ ï¿½Ï·ï¿½.
+		// ï¿½ï¿½ï¿½ï¿½Å¥ ï¿½ï¿½È¸ï¿½Ï¸é¼­ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ Ã¤ï¿½ï¿½ï¿½.
 		for (const RenderCommand& command : renderQueue)
 		{
-			// È­¸é¿¡ ±×¸± ÅØ½ºÆ®°¡ ¾øÀ¸¸é °Ç³Ê¶Ü.
+			// È­ï¿½é¿¡ ï¿½×¸ï¿½ ï¿½Ø½ï¿½Æ®ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½Ç³Ê¶ï¿½.
 			if (!command.text)
 				continue;
 
-			// ¼¼·Î ±âÁØ È­¸é ¹þ¾î³µ´ÂÁö È®ÀÎ.
+			// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ È­ï¿½ï¿½ ï¿½ï¿½ï¿½î³µï¿½ï¿½ï¿½ï¿½ È®ï¿½ï¿½.
 			if (command.position.y < 0 || command.position.y >= screenSize.y)
 				continue;
 
-			// È­¸é¿¡ ±×¸± ¹®ÀÚ¿­ ±æÀÌ
+			// È­ï¿½é¿¡ ï¿½×¸ï¿½ ï¿½ï¿½ï¿½Ú¿ï¿½ ï¿½ï¿½ï¿½ï¿½
 			const int length = static_cast<int>(strlen(command.text));
 
-			// ¾È±×·Áµµ µÇ¸é °Ç³Ê¶Ü.
+			// ï¿½È±×·ï¿½ï¿½ï¿½ ï¿½Ç¸ï¿½ ï¿½Ç³Ê¶ï¿½.
 			if (length <= 0)
 			{
 				continue;
 			}
 
-			// xÁÂÇ¥ ±âÁØÀ¸·Î È­¸é¿¡¼­ ¹þ¾î³µ´ÂÁö È®ÀÎ.
-			const int startX = command.position.x;
-			const int endX = command.position.x + length - 1;
+			// xï¿½ï¿½Ç¥ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ È­ï¿½é¿¡ï¿½ï¿½ ï¿½ï¿½ï¿½î³µï¿½ï¿½ï¿½ï¿½ È®ï¿½ï¿½.
+			const int startX = static_cast<int>(command.position.x);
+			const int endX = static_cast<int>(command.position.x) + length - 1;
 
-			if (endX < 0 || startX >= screenSize.x)
+			if (endX < 0 || startX >= static_cast<int>(screenSize.x))
 			{
 				continue;
 			}
 
-			// ½ÃÀÛ ÀÎµ¦½º¿Í ³¡ ÀÎµ¦½º ±¸ÇÏ±â.
+			// ï¿½ï¿½ï¿½ï¿½ ï¿½Îµï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ ï¿½Îµï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½Ï±ï¿½.
 			const int visibleStart = startX < 0 ? 0 : startX;
-			const int visibleEnd = endX >= screenSize.x ? screenSize.x - 1 : endX;
+			const int visibleEnd = endX >= static_cast<int>(screenSize.x) ? static_cast<int>(screenSize.x) - 1 : endX;
 
-			// ¹®ÀÚ¿­ ¼³Á¤.
+			// ï¿½ï¿½ï¿½Ú¿ï¿½ ï¿½ï¿½ï¿½ï¿½.
 			for (int x = visibleStart; x <= visibleEnd; x++)
 			{
-				// ¹®ÀÚ¿­ ¾ÈÀÇ ¹®ÀÚ ÀÎµ¦½º.
+				// ï¿½ï¿½ï¿½Ú¿ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½Îµï¿½ï¿½ï¿½.
 				const int sourceIndex = x - startX;
 
-				// ÇÁ·¹ÀÓ  (2Â÷¿ø ¹®ÀÚ ¹è¿­)ÀÎµ¦½º.
+				// ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½  (2ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½è¿­)ï¿½Îµï¿½ï¿½ï¿½.
 				const int index = (command.position.y * screenSize.x) + x;
 
-				// ±×¸®±â ¿ì¼±¼øÀ§ ºñ±³.
+				// ï¿½×¸ï¿½ï¿½ï¿½ ï¿½ì¼±ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½.
 				if (frame->sortingOrderArray[index] > command.sortingOrder)
 					continue;
 
-				// µ¥ÀÌÅÍ ±â·Ï.
+				// ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½.
 				frame->charInfoArray[index].Char.AsciiChar = command.text[sourceIndex];
 				frame->charInfoArray[index].Attributes = (WORD)command.color;
 
-				// ¿ì¼±¼øÀ§ ¾÷µ¥ÀÌÆ®.
+				// ï¿½ì¼±ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½Æ®.
 				frame->sortingOrderArray[index] = command.sortingOrder;
 			}
 		}
 
-		// ±×¸®±â.
+		// ï¿½×¸ï¿½ï¿½ï¿½.
 		GetCurrentBuffer()->Draw(frame->charInfoArray);
 
-		// ¹öÆÛ ±³È¯.
+		// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½È¯.
 		Present();
 
-		// ·»´õ Å¥ ºñ¿ì±â.
+		// ï¿½ï¿½ï¿½ï¿½ Å¥ ï¿½ï¿½ï¿½ï¿½.
 		renderQueue.clear();
 	}
 
 	void Renderer::Clear()
 	{
-		// È­¸é Áö¿ì±â.
-		// 1. ÇÁ·¹ÀÓ(2Â÷¿ø ¹è¿­ µ¥ÀÌÅÍ) Áö¿ì±â.
+		// È­ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½.
+		// 1. ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½(2ï¿½ï¿½ï¿½ï¿½ ï¿½è¿­ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½) ï¿½ï¿½ï¿½ï¿½ï¿½.
 		frame->Clear(screenSize);
 
-		// 2. ÄÜ¼Ö ¹öÆÛ Áö¿ì±â.
+		// 2. ï¿½Ü¼ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½.
 		GetCurrentBuffer()->Clear();
 	}
 
 	void Renderer::Present()
 	{
-		// ¹öÆÛ ±³È¯.
+		// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½È¯.
 		SetConsoleActiveScreenBuffer(GetCurrentBuffer()->GetBuffer());
 
-		// ÀÎµ¦½º ±³Ã¼.
+		// ï¿½Îµï¿½ï¿½ï¿½ ï¿½ï¿½Ã¼.
 		currentBufferIndex = 1 - currentBufferIndex;
 	}
 
 
 	void Renderer::Submit(const char* text, const Vector2& position, Color color, int sortingOrder)
 	{
-		// ·»´õ µ¥ÀÌÅÍ »ý¼º ÈÄ Å¥¿¡ Ãß°¡.
+		// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ Å¥ï¿½ï¿½ ï¿½ß°ï¿½.
 		RenderCommand command = {};
 		command.text = text;
 		command.position = position;
